refactor: move flowchart api function to hooks/useApi
flowchart api
